Продолжение переноса routers в dashboard module
