+++ app/modules/dashboard/__init__.py
@@ -1,5 +1,5 @@
 from .dashboard_routers import dashboard_bp
 
-__export__ = [
-    dashboard_bp
+__all__ = [
+    "dashboard_bp"
 ]
