+++ app/modules/news/__init__.py
@@ -1,8 +1,5 @@
 
 
-
-
-
 from .news_model import NewsModel
 from .news_routers import news_bp
