+++ app/routes/dashboard.py
@@ -6,7 +6,7 @@ from functools import wraps
 from app.models.warehouse import Reservation, get_warehouse_products
 from app.routes.warehouse import warehouse_bp
 
-dashboard_bp = Blueprint('dashboard', __name__)
+# dashboard_bp = Blueprint('dashboard', __name__)
 
 
 # Функция-декоратор для проверки авторизации
@@ -177,125 +177,3 @@ def manage_users():
     return render_template('admin/manage_users_page.html')
 
 
-# Вспомогательные функции
-def get_user_data(user_id):
-    """
-    Получает данные пользователя из базы данных.
-    В демонстрационных целях возвращает тестовые данные.
-    """
-    test_data = {
-        'admin': {
-            'name': 'Николай Владеев',
-            'company': 'Vision Trend',
-        },
-        'support': {
-            'name': 'Анна Поддержкина',
-            'company': 'Vision Trend',
-        },
-        'store': {
-            'name': 'Сергей Магазинов',
-            'company': 'Оптика Плюс',
-        },
-        'branch': {
-            'name': 'Иван Филиалов',
-            'company': 'Оптика Плюс',
-            'location': 'ТЦ Мега'
-        }
-    }
-
-    role = session.get('role')
-    return test_data.get(role, {})
-
-
-def get_recent_orders(user_id, role):
-    """
-    Получает список недавних заказов пользователя.
-    В демонстрационных целях возвращает тестовые данные.
-    """
-    # Демонстрационные заказы
-    orders = [
-        {'id': '4311', 'status_text': 'Отправлен', 'status_class': 'sent', 'customer': 'Оптика Плюс'},
-        {'id': '4299', 'status_text': 'Доставлен', 'status_class': 'delivered', 'customer': 'Линзы и очки'},
-        {'id': '4287', 'status_text': 'В обработке', 'status_class': 'processing', 'customer': 'ОчкиМаркет'}
-    ]
-
-    return orders
-
-
-def get_order_history(user_id, role):
-    """
-    Получает полную историю заказов пользователя.
-    В демонстрационных целях возвращает тестовые данные.
-    """
-    # Демонстрационные заказы для истории
-    orders = [
-        {
-            'id': '4311',
-            'status_text': 'Отправлен',
-            'status_class': 'sent',
-            'customer': 'Оптика Плюс',
-            'date': '10.03.2025',
-            'amount': '150,000'
-        },
-        {
-            'id': '4299',
-            'status_text': 'Доставлен',
-            'status_class': 'delivered',
-            'customer': 'Линзы и очки',
-            'date': '25.02.2025',
-            'amount': '78,500'
-        },
-        {
-            'id': '4287',
-            'status_text': 'В обработке',
-            'status_class': 'processing',
-            'customer': 'ОчкиМаркет',
-            'date': '15.02.2025',
-            'amount': '95,200'
-        },
-        {
-            'id': '4265',
-            'status_text': 'Одобрен',
-            'status_class': 'approved',
-            'customer': 'Оптика Центр',
-            'date': '01.02.2025',
-            'amount': '120,000'
-        },
-        {
-            'id': '4243',
-            'status_text': 'Отклонен',
-            'status_class': 'rejected',
-            'customer': 'СуперОчки',
-            'date': '20.01.2025',
-            'amount': '65,000'
-        }
-    ]
-
-    return orders
-
-
-def get_order_details(order_id, user_id, role):
-    """
-    Получает детальную информацию о конкретном заказе.
-    В демонстрационных целях проверяет, есть ли заказ с таким ID
-    в тестовых данных.
-    """
-    orders = get_order_history(user_id, role)
-    for order in orders:
-        if order['id'] == order_id:
-            return order
-    return None
-
-
-def get_role_title(role):
-    """
-    Возвращает название роли для отображения на странице.
-    """
-    role_titles = {
-        'admin': 'Владелец',
-        'support': 'Сотрудник',
-        'store': 'Партнёр',
-        'branch': 'Отдел партнёра'
-    }
-
-    return role_titles.get(role, 'Пользователь')
