+++ app/templates/news/add_news.html
@@ -1,7 +1,102 @@
-
-<form action="{{ url_for('news.add_news') }}" method="post">
-    <input type="text" name="title" placeholder="Title">
-    <textarea name="description" placeholder="Description"></textarea>
-    <button type="submit">Add News</button>
-    <a href="{{ url_for('news.all_news') }}">Назад</a>
-</form>
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Добавить новость</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            background-color: #1e1e1e;
+            color: #000;
+            margin: 0;
+            padding: 0;
+        }
+        .container {
+            max-width: 700px;
+            margin: 0 auto;
+            padding: 0;
+            background-color: white;
+            border-radius: 20px;
+            min-height: 100vh;
+        }
+        .header {
+            display: flex;
+            align-items: center;
+            padding: 15px 20px;
+            border-bottom: 1px solid #eee;
+        }
+        .back-btn {
+            color: #000;
+            text-decoration: none;
+            font-size: 24px;
+            margin-right: 20px;
+        }
+        .title {
+            font-size: 18px;
+            font-weight: bold;
+        }
+        .form-container {
+            padding: 20px;
+        }
+        form {
+            display: flex;
+            flex-direction: column;
+        }
+        .form-group {
+            margin-bottom: 20px;
+        }
+        label {
+            font-size: 14px;
+            font-weight: bold;
+            margin-bottom: 5px;
+            display: block;
+            color: #666;
+        }
+        input, textarea {
+            width: 100%;
+            padding: 10px;
+            border: 1px solid #eee;
+            border-radius: 5px;
+            font-size: 16px;
+            resize: vertical;
+        }
+        textarea {
+            min-height: 200px;
+        }
+        button {
+            background-color: #007bff;
+            color: white;
+            border: none;
+            padding: 12px;
+            border-radius: 5px;
+            font-size: 16px;
+            cursor: pointer;
+            margin-top: 10px;
+            display: none;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <a href="{{ url_for('news.all_news') }}" class="back-btn">←</a>
+            <div class="title">Добавить новость</div>
+        </div>
+        
+        <div class="form-container">
+            <form action="{{ url_for('news.add_news') }}" method="post">
+                <div class="form-group">
+                    <label>Заголовок</label>
+                    <input type="text" name="title" placeholder="">
+                </div>
+                <div class="form-group">
+                    <label>Текст</label>
+                    <textarea name="description" placeholder=""></textarea>
+                </div>
+                <button type="submit">Add News</button>
+            </form>
+        </div>
+    </div>
+</body>
+</html>
