+++ app/templates/news/detailed_news.html
@@ -1,10 +1,58 @@
-Hello detailed news 
-
-
-<h1>{{ news.title }}</h1>
-<p>{{ news.description }}</p>
-<p>{{ news.created_date }}</p>
-<p>{{ news.update_date }}</p>
-
-
-<a href="{{ url_for('news.all_news') }}">Назад</a>
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{ news.title }}</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            background-color: #1e1e1e;
+            color: #000;
+            margin: 0;
+            padding: 0;
+        }
+        .container {
+            max-width: 700px;
+            margin: 0 auto;
+            padding: 20px;
+            background-color: white;
+            border-radius: 20px;
+            min-height: 100vh;
+        }
+        .back-btn {
+            color: #000;
+            text-decoration: none;
+            font-size: 24px;
+            display: block;
+            margin-bottom: 20px;
+        }
+        .news-title {
+            font-size: 24px;
+            font-weight: bold;
+            margin-bottom: 15px;
+        }
+        .news-description {
+            font-size: 16px;
+            line-height: 1.5;
+            margin-bottom: 20px;
+        }
+        .date {
+            font-size: 14px;
+            color: #666;
+            display: block;
+            margin-top: 20px;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <a href="{{ url_for('news.all_news') }}" class="back-btn">←</a>
+        
+        <h1 class="news-title">{{ news.title }}</h1>
+        <div class="news-description">{{ news.description }}</div>
+        <span class="date">Создано: {{ news.created_date }}</span>
+        <span class="date">Обновлено: {{ news.update_date }}</span>
+    </div>
+</body>
+</html>
